% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{legis_get_persons}
\alias{legis_get_persons}
\alias{legis_get_person}
\title{Get persons from a Legistar site}
\usage{
legis_get_persons(
  ...,
  client = getOption("legistarapi.client"),
  simplifyVector = TRUE,
  error_call = caller_env()
)

legis_get_person(
  person_id,
  ...,
  client = getOption("legistarapi.client"),
  simplifyVector = TRUE,
  error_call = caller_env()
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=legistar]{legistar}}
  \describe{
    \item{\code{top,skip,select,filter,orderby}}{Passed to internal \code{\link[=req_odata_query]{req_odata_query()}}
function.}
    \item{\code{direction}}{Default \code{NULL}. Use \code{"asc"} for ascending or \code{"desc"} for
descending sort order based on \code{orderby} argument. Ignored if \code{orderby} is
not supplied.}
    \item{\code{count,inlinecount}}{Currently unsupported.}
    \item{\code{n_max}}{Maximum number of records to return. Currently unsupported.}
    \item{\code{.multi}}{Controls what happens when an element of \code{...} is a vector
containing multiple values:
\itemize{
\item \code{"error"}, the default, throws an error.
\item \code{"comma"}, separates values with a \verb{,}, e.g. \verb{?x=1,2}.
\item \code{"pipe"}, separates values with a \code{|}, e.g. \code{?x=1|2}.
\item \code{"explode"}, turns each element into its own parameter, e.g. \code{?x=1&x=2}.
}

If none of these functions work, you can alternatively supply a function
that takes a character vector and returns a string.}
  }}

\item{client}{String with Legistar client name. Defaults to
\code{getOption("legistarapi.client")}.}

\item{simplifyVector}{Should JSON arrays containing only primitives (i.e.
booleans, numbers, and strings) be caused to atomic vectors?}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{person_id}{Person ID. Required for \code{\link[=legis_get_person]{legis_get_person()}}.}
}
\description{
See \href{https://support.granicus.com/s/article/Adding-a-New-Person}{Adding a New Person} for more
information on persons in Legistar.
}
