% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legistar.R
\name{legistar}
\alias{legistar}
\title{Access data from the Legistar Web API}
\usage{
legistar(
  ...,
  top = NULL,
  skip = NULL,
  .multi = "pipe",
  template = "actions",
  client = getOption("legistarapi.client"),
  simplifyVector = TRUE,
  error_call = caller_env()
)
}
\arguments{
\item{...}{Template variables.}

\item{top, skip}{Passed to \code{\link[httr2:req_url]{httr2::req_url_query()}}}

\item{.multi}{Controls what happens when an element of \code{...} is a vector
containing multiple values:
\itemize{
\item \code{"error"}, the default, throws an error.
\item \code{"comma"}, separates values with a \verb{,}, e.g. \verb{?x=1,2}.
\item \code{"pipe"}, separates values with a \code{|}, e.g. \code{?x=1|2}.
\item \code{"explode"}, turns each element into its own parameter, e.g. \code{?x=1&x=2}.
}

If none of these functions work, you can alternatively supply a function
that takes a character vector and returns a string.}

\item{template}{A template string which consists of a optional HTTP method
and a path containing variables labelled like either \verb{:foo} or \code{{foo}}.}

\item{client}{String with Legistar client name}

\item{simplifyVector}{Should JSON arrays containing only primitives (i.e.
booleans, numbers, and strings) be caused to atomic vectors?}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Review the \href{https://webapi.legistar.com/}{Legistar Web API} documentation for
more information.
}
\examples{
legistar(template = "persons", client = "seattle")

legistar(template = "bodies", client = "baltimore")

}
\seealso{
\code{\link[=legistar_template]{legistar_template()}}
}
