% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{legistar_get}
\alias{legistar_get}
\alias{legistar_get_actions}
\alias{legistar_get_bodies}
\alias{legistar_get_persons}
\alias{legistar_get_body_types}
\alias{legistar_get_code_sections}
\alias{legistar_get_events}
\alias{legistar_get_indexes}
\alias{legistar_get_matter_requesters}
\alias{legistar_get_matters}
\alias{legistar_get_matter_statuses}
\alias{legistar_get_matter_types}
\alias{legistar_get_office_records}
\alias{legistar_get_vote_types}
\title{Get Legistar records}
\usage{
legistar_get_actions(
  client = getOption("legistarapi.client"),
  simplifyVector = TRUE,
  error_call = caller_env()
)

legistar_get_persons(
  client = getOption("legistarapi.client"),
  simplifyVector = TRUE,
  error_call = caller_env()
)

legistar_get_bodies(
  client = getOption("legistarapi.client"),
  simplifyVector = TRUE,
  error_call = caller_env()
)

legistar_get_body_types(
  client = getOption("legistarapi.client"),
  simplifyVector = TRUE,
  error_call = caller_env()
)

legistar_get_code_sections(
  client = getOption("legistarapi.client"),
  simplifyVector = TRUE,
  error_call = caller_env()
)

legistar_get_events(
  client = getOption("legistarapi.client"),
  simplifyVector = TRUE,
  error_call = caller_env()
)

legistar_get_indexes(
  client = getOption("legistarapi.client"),
  simplifyVector = TRUE,
  error_call = caller_env()
)

legistar_get_matter_requesters(
  client = getOption("legistarapi.client"),
  simplifyVector = TRUE,
  error_call = caller_env()
)

legistar_get_matter_requesters(
  client = getOption("legistarapi.client"),
  simplifyVector = TRUE,
  error_call = caller_env()
)

legistar_get_matters(
  client = getOption("legistarapi.client"),
  simplifyVector = TRUE,
  error_call = caller_env()
)

legistar_get_matter_statuses(
  client = getOption("legistarapi.client"),
  simplifyVector = TRUE,
  error_call = caller_env()
)

legistar_get_matter_types(
  client = getOption("legistarapi.client"),
  simplifyVector = TRUE,
  error_call = caller_env()
)

legistar_get_office_records(
  client = getOption("legistarapi.client"),
  simplifyVector = TRUE,
  error_call = caller_env()
)

legistar_get_vote_types(
  client = getOption("legistarapi.client"),
  simplifyVector = TRUE,
  error_call = caller_env()
)
}
\arguments{
\item{client}{String with Legistar client name}

\item{simplifyVector}{Should JSON arrays containing only primitives (i.e.
booleans, numbers, and strings) be caused to atomic vectors?}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Get Legistar records
}
