% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{legistar_methods}
\alias{legistar_methods}
\title{Legistar API Methods Reference}
\format{
A data frame with 124 rows and 8 variables:
\describe{
\item{\code{type}}{API method type}
\item{\code{method}}{HTTP Request Method}
\item{\code{template}}{Template string}
\item{\code{url}}{Help URL}
\item{\code{description}}{API method description}
\item{\code{response}}{Expected API method response}
\item{\code{modelName}}{Corresponding modelName for response}
\item{\code{nm}}{Snake case method name}
}
}
\source{
https://webapi.legistar.com/
}
\usage{
legistar_methods
}
\description{
A data frame with reference information from the Legistar Web API
documentation: \url{https://webapi.legistar.com/}. Organized in a Google Sheet
for use with this package:
\url{https://docs.google.com/spreadsheets/d/1Py-Oqyb8kW13x0_BjGJS-fw1EdZzqBeIsjxvQvBV5dM/edit?usp=sharing}
Last updated 2024-05-12.
}
\keyword{datasets}
