% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{legistar_template}
\alias{legistar_template}
\title{Set Legistar Web API URL GET Template}
\usage{
legistar_template(template, method = "GET", version = 1)
}
\arguments{
\item{template}{Template string corresponding to a template endpoint, e.g.
\code{actions = "{client}/Actions"}, \code{bodies = "{client}/Bodies"},
or \code{code_sections = "{client}/CodeSections"}}
}
\description{
Review the \href{https://webapi.legistar.com/}{Legistar Web API} documentation for
more information.
}
\keyword{internal}
