% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legistar.R
\name{legistar_url_id}
\alias{legistar_url_id}
\title{Parse a Legistar URL and return an ID query parameter}
\usage{
legistar_url_id(url)
}
\description{
Parse a Legistar URL and return an ID query parameter
}
\examples{

url <- "https://baltimore.legistar.com/LegislationDetail.aspx?ID=6555833&GUID=693FE27B-A165-480F-AD19-56784438224B&Options=ID|Text|&Search="

legistar_url_id(url)

}
\keyword{internal}
