% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legistarapi-package.R
\docType{package}
\name{legistarapi-package}
\alias{legistarapi}
\alias{legistarapi-package}
\title{legistarapi: Access the Legistar Web API}
\description{
Access public data using the Legistar Web API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://elipousson.github.io/legistarapi/}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

}
\keyword{internal}
