% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legistar.R
\name{req_legistar}
\alias{req_legistar}
\title{Create Legistar Web API Request with template}
\usage{
req_legistar(
  template = "actions",
  ...,
  client = getOption("legistarapi.client")
)
}
\arguments{
\item{template}{A template string which consists of a optional HTTP method
and a path containing variables labelled like either \verb{:foo} or \code{{foo}}.}

\item{...}{Template variables.}
}
\description{
Create Legistar Web API Request with template
}
\keyword{internal}
