% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legistar.R
\name{req_odata_query}
\alias{req_odata_query}
\title{Create a ODATA query using \code{httr2::req_url_query}}
\usage{
req_odata_query(
  .req,
  ...,
  top = NULL,
  skip = NULL,
  filter = NULL,
  select = NULL,
  orderby = NULL,
  direction = NULL,
  count = NULL,
  inlinecount = NULL,
  .multi = "comma",
  error_call = caller_env()
)
}
\description{
\code{req_odata_query()} provides incomplete support for the ODATA query
conventions through the \code{top}, \code{skip}, \code{filter}, \code{select}, \code{orderby}, and
\code{inlinecount} parameters. Working examples using the Legistar API
}
\details{
Created using the \href{https://www.odata.org/documentation/odata-version-3-0/odata-version-3-0-core-protocol/}{OData Version 3.0 Core Protocol}
linked from the Legistar API documentation for reference.
}
\keyword{internal}
