% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legistar.R
\name{request_legistar}
\alias{request_legistar}
\title{Create Legistar Web API Request}
\usage{
request_legistar(base_url = "https://webapi.legistar.com")
}
\arguments{
\item{base_url}{Base URL for request.}
}
\description{
Create Legistar Web API Request
}
\keyword{internal}
