% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legistar.R
\name{resp_legistar}
\alias{resp_legistar}
\title{Extract Legistar Web API Response Body}
\usage{
resp_legistar(
  resp,
  ...,
  simplifyVector = FALSE,
  error_call = rlang::caller_env()
)
}
\arguments{
\item{resp}{A response object.}

\item{...}{Other arguments passed on to \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}} and
\code{\link[xml2:read_xml]{xml2::read_xml()}} respectively.}

\item{simplifyVector}{Should JSON arrays containing only primitives (i.e.
booleans, numbers, and strings) be caused to atomic vectors?}
}
\description{
Extract Legistar Web API Response Body
}
\keyword{internal}
