% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legistar.R
\name{set_legistar_client}
\alias{set_legistar_client}
\alias{get_legistar_client}
\title{Set and get an option for \code{legistarapi.client}}
\usage{
set_legistar_client(client)

get_legistar_client(default = NULL)
}
\arguments{
\item{client}{String with client name}

\item{default}{if the specified option is not set in the options list,
    this value is returned.  This facilitates retrieving an option and
    checking whether it is set and setting it separately if not.}
}
\description{
Set and get an option for \code{legistarapi.client}
}
\keyword{internal}
