% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legistar.R
\name{legistar}
\alias{legistar}
\title{Access data from the Legistar Web API}
\usage{
legistar(
  ...,
  template = "actions",
  client = getOption("legistarapi.client"),
  top = NULL,
  skip = NULL,
  select = NULL,
  filter = NULL,
  orderby = NULL,
  direction = NULL,
  count = NULL,
  inlinecount = NULL,
  .multi = "pipe",
  simplifyVector = TRUE,
  n_max = NULL,
  error_call = caller_env()
)
}
\arguments{
\item{...}{Template variables.}

\item{template}{A template string which consists of a optional HTTP method
and a path containing variables labelled like either \verb{:foo} or \code{{foo}}.}

\item{client}{String with Legistar client name. Defaults to
\code{getOption("legistarapi.client")}.}

\item{top, skip, select, filter, orderby}{Passed to internal \code{\link[=req_odata_query]{req_odata_query()}}
function.}

\item{count, inlinecount}{Currently unsupported.}

\item{.multi}{Controls what happens when an element of \code{...} is a vector
containing multiple values:
\itemize{
\item \code{"error"}, the default, throws an error.
\item \code{"comma"}, separates values with a \verb{,}, e.g. \verb{?x=1,2}.
\item \code{"pipe"}, separates values with a \code{|}, e.g. \code{?x=1|2}.
\item \code{"explode"}, turns each element into its own parameter, e.g. \code{?x=1&x=2}.
}

If none of these functions work, you can alternatively supply a function
that takes a character vector and returns a string.}

\item{simplifyVector}{Should JSON arrays containing only primitives (i.e.
booleans, numbers, and strings) be caused to atomic vectors?}

\item{n_max}{Maximum number of records to return. Currently unsupported.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{order}{Default \code{NULL}. Use \code{"asc"} for ascending or \code{"desc"} for
descending sort order based on \code{orderby} argument. Ignored if \code{orderby} is
not supplied.}
}
\description{
Review the \href{https://webapi.legistar.com/}{Legistar Web API} documentation for
more information. This function is a work in progress. As of May 2024, it
returns all results for any query calling itself recursively.
}
\examples{
legistar(template = "persons", client = "seattle", top = 5)

legistar(template = "bodies", client = "baltimore", top = 5)

}
\seealso{
\code{\link[=legistar_template]{legistar_template()}}
}
