% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{legistar_clients}
\alias{legistar_clients}
\title{Legistar Client Reference}
\format{
A data frame with 286 rows and 2 variables:
\describe{
\item{\code{client}}{Client string}
\item{\code{api}}{logical If \code{TRUE}, client returned an API response}
}
}
\usage{
legistar_clients
}
\description{
A data frame of identified Legistar client sites tested to check if client
responded to an API request with \code{\link[=legis_get_bodies]{legis_get_bodies()}}. Last updated
2024-05-12.
}
\keyword{datasets}
